pyproject.toml
[tool.poetry]
name = "stock_II"
version = "0.1.0"
description = "Stock project for trading & visualization"
authors = ["Leon Schmidt <leon.schmidt@charite.de>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
matplotlib = "^3.3.3"
scipy = "^1.5.4"
pandas = "^1.1.5"
numpy = "^1.19.4"
black = "^20.8b1"
toml = "^0.10.2"
pytest = "^6.2.1"
psycopg2 = "^2.8.6"
docker-compose = "^1.27.4"
Flask = "^1.1.2"
Flask-Jsonpify = "^1.5.0"
Flask-RESTful = "^0.3.8"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.6.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
